Split out into only darwin and windows
