+++ osquery/tables/applications/windows/carbon_black.cpp
@@ -11,15 +11,11 @@
 #include <osquery/tables.h>
 #include <osquery/core.h>
 #include <osquery/filesystem.h>
-#ifdef WIN32
 #include <osquery/tables/system/windows/registry.h>
-#endif
 
 #include <boost/filesystem.hpp>
 #include <boost/property_tree/ptree.hpp>
-#include <boost/property_tree/ini_parser.hpp>
 #include <boost/algorithm/string/replace.hpp>
-#include <boost/algorithm/string/predicate.hpp>
 
 namespace fs = boost::filesystem;
 
@@ -28,65 +24,8 @@ namespace tables {
 
 // Carbon Black registry path 
 #define kCbRegLoc "SOFTWARE\\CarbonBlack\\config"
-// Path to the Carbon Black sensor id file
-#define kCbSensorIdFile "/var/lib/cb/sensor.id"
-// Path to the Carbon Black sensor settings file
-#define kCbSensorSettingsFile "/var/lib/cb/sensorsettings.ini"
 // Path to Carbon Black direcotry
-#ifdef WIN32
 #define kCbDir "C:\\Windows\\CarbonBlack\\"
-#else
-#define kCbDir "/var/lib/cb/"
-#endif
-
-
-// Get the Carbon Black sensor ID
-void getSensorId(Row& r) {
-    std::string file_contents;
-    if (!forensicReadFile(kCbSensorIdFile, file_contents).ok()) {
-        return;
-    }
-    // check to make sure we have sane data
-    if (file_contents.length() != 16) {
-        return;
-    }
-
-    unsigned int sensor_id;
-    std::string hex_sensor_id = file_contents.substr(11, 16);
-    std::stringstream converter(hex_sensor_id);
-    converter >> std::hex >> sensor_id;
-    r["sensor_id"] = INTEGER(sensor_id);
-}
-
-// Get settings of the Carbon Black sensor
-void getSensorSettings(Row& r) {
-    if (!boost::filesystem::exists(kCbSensorSettingsFile)) {
-        return;
-    }
-    boost::property_tree::ptree pt;
-    boost::property_tree::ini_parser::read_ini(kCbSensorSettingsFile, pt);
-    r["config_name"] = SQL_TEXT(pt.get<std::string>("CB.ConfigName"));
-    r["collect_store_files"] = INTEGER(pt.get<std::string>("CB.CollectStoreFiles"));
-    r["collect_module_loads"] = INTEGER(pt.get<std::string>("CB.CollectModuleLoads"));
-    r["collect_module_info"] = INTEGER(pt.get<std::string>("CB.CollectModuleInfo"));
-    r["collect_file_mods"] = INTEGER(pt.get<std::string>("CB.CollectFileMods"));
-    r["collect_reg_mods"] = INTEGER(pt.get<std::string>("CB.CollectRegMods"));
-    r["collect_net_conns"] = INTEGER(pt.get<std::string>("CB.CollectNetConns"));
-    r["collect_processes"] = INTEGER(pt.get<std::string>("CB.CollectProcesses"));
-    r["collect_cross_processes"] = INTEGER(pt.get<std::string>("CB.CollectCrossProcess"));
-    r["collect_emet_events"] = INTEGER(pt.get<std::string>("CB.CollectEmetEvents"));
-    std::string server = pt.get<std::string>("CB.SensorBackendServer");
-    boost::replace_all(server, "%3A", ":");
-    r["sensor_backend_server"] = SQL_TEXT(server);
-    r["collect_data_file_writes"] = INTEGER(0);
-    r["collect_processes"] = INTEGER(0);
-    r["collect_sensor_operations"] = INTEGER(0);
-    r["log_file_disk_quota_mb"] = INTEGER(0);
-    r["log_file_disk_quota_percentage"] = INTEGER(0);
-    r["protection_disabled"] = INTEGER(0);
-    r["collect_process_user_context"] = INTEGER(0);
-    r["sensor_ip_addr"] = SQL_TEXT("");
-}
 
 void getQueue(Row& r) {
     std::vector<std::string> files_list;
@@ -99,15 +38,12 @@ void getQueue(Row& r) {
     // Go through each file
     for (const auto& kfile : files_list) {
         fs::path file(kfile);
-        if (file.filename() == "filedata" ||
-            file.filename() == "metadata" ||
-            file.filename() == "data" ||
+        if (file.filename() == "data" ||
             file.filename() == "info.txt") {
 
             binary_queue_size += fs::file_size(kfile);
         }
-        if (file.stem() == "events" ||
-            file.filename() == "active-event.log" ||
+        if (file.filename() == "active-event.log" ||
             boost::starts_with(file.filename().c_str(), "eventlog_")) {
             event_queue_size += fs::file_size(kfile);
         }
@@ -116,7 +52,6 @@ void getQueue(Row& r) {
     r["event_queue"] = INTEGER(event_queue_size);
 }
 
-#ifdef WIN32
 void getWinSettings(Row& r) {
     QueryData results;
     queryKey("HKEY_LOCAL_MACHINE", kCbRegLoc, results);
@@ -193,18 +128,12 @@ void getWinSettings(Row& r) {
         }
     }
 }
-#endif
 
 QueryData genInfo(QueryContext &context) {
     Row r;
     QueryData results;
 
-#ifdef WIN32
-    getWinSettings(r);
-#else
-    getSensorId(r);
-    getSensorSettings(r);
-#endif
+    getSettings(r);
     getQueue(r);
     results.push_back(r);
