+++ specs/windows/cb_info.table
